【第21节】 1. 完善了事务结构 2. 仍然保留了充值bug 3. 删除了不需要的文件

+++ big-market-domain/src/main/java/com/dasi/domain/activity/model/entity/RaffleAwardEntity.java
@@ -26,7 +26,7 @@ public class RaffleAwardEntity {
     private String orderId;
 
     /** 奖品id */
-    private Integer awardId;
+    private Long awardId;
 
     /** 奖品标题 */
     private String awardName;
