+++ big-market-domain/src/main/java/com/dasi/domain/strategy/service/rule/chain/impl/RuleWeightChain.java
@@ -59,7 +59,7 @@ public class RuleWeightChain extends AbstractRuleChain {
 
         // 5. 如果匹配上积分阈值，则在当前积分阈值下抽奖
         if (matchedThreshold != null) {
-            Integer awardId = strategyLottery.getLotteryAward(strategyId, String.valueOf(matchedThreshold));
+            Long awardId = strategyLottery.getLotteryAward(strategyId, String.valueOf(matchedThreshold));
             log.info("【策略责任链】rule_weight 接管：匹配积分={}，awardId={}", matchedThreshold, awardId);
             return RuleCheckResult.builder()
                     .awardId(awardId)
