+++ big-market-domain/src/main/java/com/dasi/domain/strategy/service/rule/tree/IRuleTree.java
@@ -5,6 +5,6 @@ import com.dasi.domain.strategy.model.dto.RuleCheckResult;
 
 public interface IRuleTree {
 
-    RuleCheckResult logic(String userId, Long strategyId, Integer awardId, String ruleValue);
+    RuleCheckResult logic(String userId, Long strategyId, Long awardId, String ruleValue);
 
 }
