+++ big-market-domain/src/main/java/com/dasi/domain/strategy/service/rule/tree/impl/RuleLuckTree.java
@@ -14,8 +14,8 @@ import org.springframework.stereotype.Component;
 public class RuleLuckTree implements IRuleTree {
 
     @Override
-    public RuleCheckResult logic(String userId, Long strategyId, Integer awardId, String ruleValue) {
-        Integer luckAwardId = Integer.valueOf(ruleValue);
+    public RuleCheckResult logic(String userId, Long strategyId, Long awardId, String ruleValue) {
+        Long luckAwardId = Long.valueOf(ruleValue);
         log.info("【策略规则树】rule_luck 接管：awardId={}", luckAwardId);
         return RuleCheckResult.builder()
                 .awardId(luckAwardId)
